+++ ms_deisotope/data_source/__init__.py
@@ -31,6 +31,10 @@ from .scan import ScanProxyContext
 
 from ._compression import get_opener
 
+from .text import scan_from_csv
+from .memory import make_scan
+
+
 __all__ = [
     "MSFileLoader", "MzMLLoader",
     "MzXMLLoader", "MGFLoader",
@@ -47,5 +51,6 @@ __all__ = [
 
     "ScanDataSource", "ScanIterator", "ScanBunch",
     "ScanWindow", "RandomAccessScanSource", "ChargeNotProvided",
-    "get_opener", "ScanProxyContext",
+    "get_opener", "ScanProxyContext", 'scan_from_csv',
+    "make_scan"
 ]